+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - (Splunk) The collectd/jenkins plugin is deprecated. Please see https://plugins.jenkins.io/opentelemetry/ to enable OpenTelemetry signals. This plugin will be removed by the end of October 2025. ([#6489](https://github.com/signalfx/splunk-otel-collector/pull/6489))
 - (Splunk) The collectd/rabbitmq plugin is deprecated. Please use the [rabbitmq receiver](https://help.splunk.com/en/splunk-observability-cloud/manage-data/available-data-sources/supported-integrations-in-splunk-observability-cloud/opentelemetry-receivers/rabbitmq-receiver) instead.
   This plugin will be removed by the end of October 2025. ([#6490](https://github.com/signalfx/splunk-otel-collector/pull/6490))
+- (Splunk) The collectd/redis plugin is deprecated. Please use the [redis receiver](https://help.splunk.com/en/splunk-observability-cloud/manage-data/available-data-sources/supported-integrations-in-splunk-observability-cloud/opentelemetry-receivers/redis-receiver) instead.
+  This plugin will be removed by the end of October 2025. ([#6491](https://github.com/signalfx/splunk-otel-collector/pull/6491))
 
 ## v0.130.0